Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/unsupported-csp-referrer-directive-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'referrer'.

Referrer Policy: CSP 'referrer' directive should not be supported

CSP used to have a 'referrer' directive to set a Referrer Policy. This directive has been removed and should not be supported.

Running...


PASS Image has a referrer despite CSP 'referrer' directive
